Add files via upload
Run from the directory you want to put the files into and make sure you have read write access.  I don't do any error checking yet.  it is on my list of things to do.  This is basically command line only. I am working on a gui version.  not sure when this will be ready.